Split AbstractNode into four classes.
AttributesHolder -> TypedAttributesHolder -> SerializableAttributesHolder -> AbstractNode

AttributesHolder and TypedAttributesHolder: theoretical part of the datamodel specification
SerializableAttributesHolder: serialization part of the spec
AbstractNode: utilities for making operations on trees.
