+++ ppp_datamodel/__init__.py
@@ -1,4 +1,5 @@
 """Classes representing the data model of the Projet Pensées Profondes."""
 
+from .utils import *
 from .nodes import *
 from .communication import *
