+++ ppp_datamodel/nodes/abstractnode.py
@@ -2,6 +2,7 @@
 
 import json
 
+from ..utils.typedattributesholder import TypedAttributesHolder
 from ..log import logger
 from .. import exceptions
 
@@ -12,97 +13,16 @@ def register(cls):
     TYPE_TO_CLASS[cls._type] = cls
     return cls
 
-class AbstractNode:
-    """Base class for PPP nodes."""
-    __slots__ = ('_attributes')
-    _type = None
-    _possible_attributes = None
-    def __init__(self, *args, **attributes):
-        # Sanity checks
-        if self._type is None or self._possible_attributes is None:
-            raise TypeError('%s is an abstract class.' % self.__class__)
-        attributes.update(dict(zip(self._possible_attributes, args)))
-        self._check_attributes(attributes)
-
-        self._parse_attributes(attributes)
-        self._attributes['type'] = self.type
-
-    def _check_attributes(self, attributes, extra=None):
-        """Check if attributes given to the constructor can be used to
-        instanciate a valid node."""
-        extra = extra or ()
-        if 'type' in attributes:
-            assert attributes.pop('type') == self.type
-        unknown_keys = set(attributes) - set(self._possible_attributes) - set(extra)
-        if unknown_keys:
-            logger.warning('%s node got unknown attributes: %s' %
-                            (self._type, unknown_keys))
-
-    def _parse_attributes(self, attributes):
-        self._attributes = attributes
-
-    @property
-    def type(self):
-        """Type of the node."""
-        return self._type
-
-    def __repr__(self):
-        return '<PPP node "%s" %r>' % (self.type,
-                {x:y for (x,y) in self._attributes.items() if x != 'type'})
-
-    def __eq__(self, other):
-        """Tests equality with another abstractnode instance."""
-        if isinstance(other, dict):
-            return self.as_dict() == other
-        elif isinstance(other, AbstractNode):
-            return self._attributes == other._attributes
-        else:
-            return False
-
-    def __hash__(self):
-        return hash(frozenset(self._attributes.items()))
-
-    def fold(self, predicate):
-        """Takes a predicate and applies it to each node starting from the
-        leaves and making the return value propagate."""
-        childs = {x:y.fold(predicate) for (x,y) in self._attributes.items()
-                  if isinstance(y, AbstractNode)}
-        return predicate(self, childs)
-
-    def get(self, name, strict=True):
-        """Get an attribute of the node (read-only access)."""
-        if not isinstance(name, str) or name.startswith('_'):
-            raise AttributeError(self.__class__.__name__, name)
-        elif strict and name not in self._possible_attributes:
-            raise AttributeError('%s is not a valid attribute of %r.' %
-                                 (name, self))
-        elif name in self._attributes:
-            return self._attributes[name]
-        else:
-            raise exceptions.AttributeNotProvided(name)
-    __getattr__ = __getitem__ = get
-
-    def __setattr__(self, name, value):
-        if name.startswith('_'):
-            super().__setattr__(name, value)
-        else:
-            raise TypeError('%s\'s attributes are not settable.' %
-                    self.__class__.__name__)
-    def __delattr__(self, name):
-        if name.startswith('_'):
-            super().__delattr__(name, value)
-        else:
-            raise TypeError('%s\'s attributes are not settable.' %
-                    self.__class__.__name__)
-
-    def has(self, name):
-        """Check existence of an attribute."""
-        return name in self._attributes
-    __hasattr__ = __contains__ = has
-
+class SerializableAttributesHolder(TypedAttributesHolder):
+    """TypedAttributesHolder with methods handling serialization and
+    deserialization according to the PPP datamodel specification."""
     def as_dict(self):
         """Returns a JSON-serializeable object representing this tree."""
-        conv = lambda v: v.as_dict() if isinstance(v, AbstractNode) else v
+        def conv(v):
+            if isinstance(v, SerializableAttributesHolder):
+                return v.as_dict()
+            else:
+                return v
         return {k.replace('_', '-'): conv(v) for (k, v) in self._attributes.items()}
     def as_json(self):
         """Return a JSON dump of the object."""
@@ -143,15 +63,25 @@ class AbstractNode:
 
     @classmethod
     def deserialize_attribute(cls, key, value):
-        return AbstractNode.from_dict(value)
+        return cls.from_dict(value)
 
     @classmethod
     def _select_class(cls, data):
         return TYPE_TO_CLASS[data['type']]
 
+class AbstractNode(SerializableAttributesHolder):
+    """SerializableAttributesHolder with methods for making operations
+    on trees."""
+    def fold(self, predicate):
+        """Takes a predicate and applies it to each node starting from the
+        leaves and making the return value propagate."""
+        childs = {x:y.fold(predicate) for (x,y) in self._attributes.items()
+                  if isinstance(y, TypedAttributesHolder)}
+        return predicate(self, childs)
+
     def traverse(self, predicate):
         def wrapper(tree):
-            if isinstance(tree, AbstractNode):
+            if isinstance(tree, TypedAttributesHolder):
                 return tree.traverse(predicate)
             else:
                 return tree
